gemini/requirements.txt
streamlit==0.89.0
PyPDF2==1.26.0
langchain==0.1.0
langchain-google-genai==0.1.0
faiss-cpu==1.7.1
python-dotenv==0.19.1
